update for move2azure
